requirements.txt
# Core dependencies
python-dotenv>=1.0.0
httpx>=0.24.0                  # For HTTP providers (Cerebras, OpenAI, Claude, etc.)

# Free Cloud Provider SDKs
groq>=0.9.0                    # Groq cloud provider
google-generativeai>=0.8.0     # Google Gemini
huggingface-hub>=0.20.0        # HuggingFace Inference

# Paid Provider SDKs
openai>=1.0.0                  # OpenAI (GPT-4, GPT-3.5)
anthropic>=0.18.0              # Claude (Anthropic)

# Note: Moonshot uses httpx (OpenAI-compatible API)
# Note: OpenRouter, GitHub Models use httpx (OpenAI-compatible APIs)
# Note: Cerebras uses httpx (OpenAI-compatible API)
# Note: Local providers (Ollama, LM Studio, llama.cpp) use httpx

# Future: AutoGen integration
# pyautogen>=0.2.0             # When implementing AC3 agents
